corrige tipos de variáveis e especificador de formato para float

+++ calculo_de_media.c
@@ -1,21 +1,21 @@
 #include <stdio.h>
 
 int main() {
-    int nota1, nota2, nota3;
+    float nota1, nota2, nota3;
     float media;
 
     printf("*** Programa de Cálculo de Média ***\n");
 
     printf("Digite a sua primeira nota: \n");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite a sua segunda nota: \n");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite a sua terceira nota: \n");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
-    media = (float)(nota1 + nota2 + nota3) / 3;
+    media = (nota1 + nota2 + nota3) / 3;
 
     printf("A média é: %.1f", media);
